backend/requirements.txt
# SuperElite - AI 风光摄影智能选片工具
# 基于 One-Align 双维度评分

# =====================
# 核心依赖
# =====================
torch>=2.0.0
torchvision>=0.15.0
transformers==4.40.0
accelerate>=0.20.0
icecream>=2.1.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# =====================
# 图像处理
# =====================
Pillow>=10.0.0

# =====================
# RAW 文件处理
# =====================
rawpy>=0.18.0
imageio>=2.30.0

# =====================
# 进度显示
# =====================
tqdm>=4.60.0

# =====================
# 注意事项
# =====================
# 1. 需要安装 exiftool:
#    brew install exiftool
#
# 2. 仅支持 Apple Silicon Mac
#
# 3. One-Align 模型约 15GB，首次运行自动下载
